refactor: 세션 생성 페이지 리팩토링
