+++ frontend/src/components/sessions/create/SessionForm/ListSelectModal/CategoryTab/Category.tsx
@@ -1,4 +1,4 @@
-import useSessionFormStore from "@stores/useSessionFormStore";
+import useSessionFormStore from "@/pages/CreateSessionPage/stores/useSessionFormStore";
 
 interface Props {
   tabName: "myList" | "savedList";
@@ -11,11 +11,10 @@ const Category = ({ tabName, tabText }: Props) => {
   return (
     <button
       className={`flex flex-row gap-1
-      ${
-        tab === tabName
+      ${tab === tabName
           ? "text-bold-s text-green-500"
           : "text-medium-l text-gray-400"
-      }`}
+        }`}
       onClick={() => setTab(tabName)}
     >
       <span
