+++ frontend/src/components/sessions/create/SessionForm/ListSelectModal/QuestionList/QuestionItem.tsx
@@ -1,6 +1,6 @@
 import { GrDown, GrUp } from "react-icons/gr";
 import { ImCheckmark } from "react-icons/im";
-import useSessionFormStore from "@/stores/useSessionFormStore";
+import useSessionFormStore from "@/pages/CreateSessionPage/stores/useSessionFormStore";
 import axios from "axios";
 import { useState } from "react";
 import LoadingIndicator from "@components/common/LoadingIndicator.tsx";
@@ -105,10 +105,9 @@ const QuestionItem = ({ item }: { item: QuestionList }) => {
         </div>
         <button
           className={`flex items-center ml-auto w-10 h-10 rounded-custom-m
-            ${
-              isSelected
-                ? "bg-green-200 text-green-50"
-                : "bg-gray-300 text-gray-50"
+            ${isSelected
+              ? "bg-green-200 text-green-50"
+              : "bg-gray-300 text-gray-50"
             }`}
           onClick={() => {
             checkHandler(item.id, item.title);
