+++ frontend/src/components/sessions/create/SessionForm/QuestionListSection.tsx
@@ -1,4 +1,4 @@
-import useSessionFormStore from "@/stores/useSessionFormStore";
+import useSessionFormStore from "@/pages/CreateSessionPage/stores/useSessionFormStore";
 import SelectTitle from "@/components/common/Text/SelectTitle";
 import { MdOutlineArrowForwardIos } from "react-icons/md";
